# Improve connection pooling