--- a/tests/Conformance.Tests/ConnectionTests.cs
+++ b/tests/Conformance.Tests/ConnectionTests.cs
@@ -10,6 +10,11 @@ namespace Conformance.Tests
 		{
 		}
 
+		[Fact(Skip = "Throws MySqlException when it attempts to connect.")]
+		public override void Set_ConnectionString_throws_when_invalid()
+		{
+		}
+
 		[Fact(Skip = "Throws MySqlException when it attempts to connect, not InvalidOperationException before connecting")]
 		public override void Open_throws_when_no_connection_string()
 		{