Defer creation of `MySqlConnectionStringBuilder` and `ConnectionSettings` until it is known that a new ConnectionPool needs to be created.

```
BenchmarkDotNet=v0.10.12, OS=macOS 10.13.3 (17D47) [Darwin 17.4.0]
Intel Core i7-3740QM CPU 2.70GHz (Ivy Bridge), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=2.1.4
  [Host]    : .NET Core 2.0.5 (Framework 4.6.0.0), 64bit RyuJIT
  netcore20 : .NET Core 2.0.5 (Framework 4.6.0.0), 64bit RyuJIT
```

### Before

Method |     Mean |     Error |    StdDev |   StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
  OpenSync | 772.8 us | 15.246 us | 30.448 us | 4.350 us | 715.7 us | 749.7 us | 781.6 us | 791.1 us | 850.3 us | 1,294.0 | 0.9766 |   5.56 KB |
 OpenAsync | 799.9 us |  7.779 us |  7.277 us | 1.879 us | 790.5 us | 793.8 us | 798.0 us | 804.5 us | 816.6 us | 1,250.2 | 2.9297 |   8.05 KB |

### After reducing allocations

Method |     Mean |    Error |   StdDev |   StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
  OpenSync | 749.5 us | 14.80 us | 23.47 us | 4.085 us | 719.0 us | 728.5 us | 753.2 us | 766.2 us | 809.5 us | 1,334.3 |      - |   1.39 KB |
 OpenAsync | 835.3 us | 16.09 us | 15.05 us | 3.885 us | 814.1 us | 823.0 us | 832.3 us | 846.3 us | 863.1 us | 1,197.1 | 1.9531 |   3.89 KB |

### Adding ConcurrentDictionary

Method |     Mean |     Error |    StdDev |    StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |---------:|----------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
  OpenSync | 730.2 us | 12.010 us | 10.029 us | 2.7814 us | 711.7 us | 723.7 us | 728.7 us | 739.8 us | 748.1 us | 1,369.4 |      - |   1.39 KB |
 OpenAsync | 802.8 us |  2.294 us |  1.916 us | 0.5314 us | 800.2 us | 801.1 us | 802.4 us | 804.3 us | 806.5 us | 1,245.7 | 1.9531 |   3.89 KB |

### GetSessionAsync returns ValueTask

Method |     Mean |     Error |    StdDev |   StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |---------:|----------:|----------:|---------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
  OpenSync | 748.5 us | 14.867 us | 23.146 us | 4.092 us | 722.6 us | 730.3 us | 745.9 us | 759.0 us | 823.0 us | 1,336.0 |      - |   1.32 KB |
 OpenAsync | 747.2 us |  8.242 us |  6.882 us | 1.909 us | 737.9 us | 742.1 us | 745.2 us | 751.4 us | 763.6 us | 1,338.3 | 1.9531 |   3.91 KB |

### Add single-entry MRU cache

Method |     Mean |    Error |   StdDev |   StdErr |      Min |       Q1 |   Median |       Q3 |      Max |    Op/s |  Gen 0 | Allocated |
---------- |---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|---------:|--------:|-------:|----------:|
  OpenSync | 724.5 us | 13.57 us | 15.62 us | 3.493 us | 696.6 us | 720.3 us | 725.3 us | 732.3 us | 757.3 us | 1,380.3 |      - |   1.32 KB |
 OpenAsync | 818.8 us | 16.22 us | 15.93 us | 3.982 us | 805.9 us | 809.5 us | 811.5 us | 821.1 us | 867.7 us | 1,221.3 | 1.9531 |   3.91 KB |